document_qa_system/requirements.txt
streamlit==1.32.2
pypdf==4.1.0
sentence-transformers==2.5.1
faiss-cpu==1.8.0
google-generativeai==0.4.1
python-dotenv==1.0.1
numpy<2.0
